notepad requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
pypdf
python-docx
email
pinecone-client
tiktoken
